refactor(image.ts)
